+++ examples/index.php
@@ -1,5 +1,0 @@
-<h2>MediaLab API examples</h2>
-<ul>
-  <li><a href="user.php">Retrieving user info.</a></li>
-  <li><a href="upload.php">Upload.</a></li>
-</ul>