+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [2.0.0] - 2016-01-01
+- Updating league/oauth2-client package to 1.1, now requiring PHP 5.5.
+- Adding Media service.
+
 ## [1.0.0] - 2015-12-24
-- Initial release
+- Initial release (PHP 5.3 compatible).